Okay... I changed everything
Now going to use visibility to hide components when they are not
supposed to be shown.  That way they will remain on the Dom, and the
page height will be the same, so scrolling will work.

Still have to finish the visibility stuff on each individual component
and calculate the height on each component to switch to the next.  But
I think this way will actually work.  Need a break, this is still broke
as heck right now, just warning you future Marla.

Still wish I knew why setting window.scrollY = 0 on componentDidMount
would cause it to be set to 0 forever.  The DOM is a mysterious place…